delete all.

+++ src/main/webapp/public/lib.js
@@ -4,7 +4,13 @@ $(document).ready(function () {
     }).then(function (data) {
         let content = "<tbody>";
 
-        content += fillTable(data);
+        if (data != null && data.length > 0) {
+            content += fillTable(data);
+        } else {
+            content += '<tr>';
+            content += '<td colspan="6" align="center">No records</td>';
+            content += '</tr>';
+        }
 
         content += "</tbody>";
         $('#elements').append(content);
@@ -27,11 +33,24 @@ function fillTable(data) {
 }
 
 function deleteElement(e) {
-    const elementId = $(e).data('id');
-    $.ajax({
-        url: "/ws-app/j2ee/v1/elements/" + elementId,
-        type: 'DELETE'
-    }).then(function (data) {
-        location.reload();
-    });
+    if (confirm("Are you sure?")) {
+        const elementId = $(e).data('id');
+        $.ajax({
+            url: "/ws-app/j2ee/v1/elements/" + elementId,
+            type: 'DELETE'
+        }).then(function (data) {
+            location.reload();
+        });
+    }
+}
+
+function deleteElements() {
+    if (confirm("Are you sure?")) {
+        $.ajax({
+            url: "/ws-app/j2ee/v1/elements",
+            type: 'DELETE'
+        }).then(function (data) {
+            location.reload();
+        });
+    }
 }
